requirements.txt
ansible-base==2.10.8
apache-libcloud==3.2.0
argcomplete==1.8.1
attrs>=22.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.6.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.1.0
fail2ban==0.11.2
Flask==3.0.0
gunicorn==21.2.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iotop==0.6
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema>=4.10.0
keyring>=24.0.0
lazr.uri==1.0.6
lockfile==0.12.2
MarkupSafe==2.1.3
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
packaging==21.3
pexpect==4.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodomex==3.11.0
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pykerberos==1.1.14
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
pywinrm==0.3.0
PyYAML==5.4.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
selinux==3.3
service-identity==18.1.0
simplejson==3.17.6
six==1.16.0
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.0
xmltodict==0.12.0
zipp==1.0.0
zope.interface==5.4.0
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
